Refactored BLE transport to use new FidoDevice abstraction
