+++ src/proto/ctap1/mod.rs
@@ -1,13 +1,13 @@
-mod manager;
+mod model;
 mod protocol;
 
 pub mod apdu;
 
-pub use self::protocol::Ctap1RegisteredKey;
-pub use self::protocol::Ctap1Transport;
-pub use self::protocol::Ctap1Version;
-pub use self::protocol::{Ctap1RegisterRequest, Ctap1RegisterResponse};
-pub use self::protocol::{Ctap1SignRequest, Ctap1SignResponse};
-pub use self::protocol::{Ctap1VersionRequest, Ctap1VersionResponse};
+pub use self::model::Ctap1RegisteredKey;
+pub use self::model::Ctap1Transport;
+pub use self::model::Ctap1Version;
+pub use self::model::{Ctap1RegisterRequest, Ctap1RegisterResponse};
+pub use self::model::{Ctap1SignRequest, Ctap1SignResponse};
+pub use self::model::{Ctap1VersionRequest, Ctap1VersionResponse};
 
-pub use self::manager::{Ctap1, Ctap1Manager};
+pub use self::protocol::{Ctap1, Ctap1Protocol};
