+++ src/proto/ctap2/mod.rs
@@ -3,16 +3,16 @@ extern crate async_trait;
 pub mod cbor;
 
 mod downgrade;
-mod manager;
+mod model;
 mod protocol;
 
 pub use downgrade::Ctap2DowngradeCheck;
-pub use manager::{Ctap2, Ctap2Manager};
-pub use protocol::Ctap2GetInfoResponse;
-pub use protocol::{
+pub use model::Ctap2GetInfoResponse;
+pub use model::{
     Ctap2COSEAlgorithmIdentifier, Ctap2CommandCode, Ctap2CredentialType,
     Ctap2PublicKeyCredentialDescriptor, Ctap2PublicKeyCredentialRpEntity,
     Ctap2PublicKeyCredentialType, Ctap2PublicKeyCredentialUserEntity, Ctap2Transport,
 };
-pub use protocol::{Ctap2GetAssertionRequest, Ctap2GetAssertionResponse};
-pub use protocol::{Ctap2MakeCredentialRequest, Ctap2MakeCredentialResponse};
+pub use model::{Ctap2GetAssertionRequest, Ctap2GetAssertionResponse};
+pub use model::{Ctap2MakeCredentialRequest, Ctap2MakeCredentialResponse};
+pub use protocol::{Ctap2, Ctap2Protocol};
