+++ src/transport/hid/mod.rs
@@ -1,3 +1,5 @@
 pub mod device;
 pub mod framing;
 pub mod init;
+
+pub use device::{list_devices, HidFidoDevice};
